Collision resolution tweak
Resolution of a collision will only change the entity's velocity vector if the dot product of the velocity with the collision manifold is a negative value (that is, the angle between vectors is greater than 90 degrees).

This fixed an issue where ascending up the side of a tile would cause entities to "suck" onto the top of the tile when reached.  Entities will now follow the correct trajectory up the side of a tile.